+++ Sources/ProjectDescription/TuistConfig.swift
@@ -4,10 +4,8 @@ import Foundation
 public struct TuistConfig: Codable, Equatable {
     /// Contains options related to the project generation.
     ///
-    /// - generateManifestElement: When passed, Tuist generates the projects, targets and schemes to compile the project manifest.
     /// - xcodeProjectName(TemplateString): When passed, Tuist generates the project with the specific name on disk instead of using the project name.
     public enum GenerationOptions: Encodable, Decodable, Equatable {
-        case generateManifest
         case xcodeProjectName(TemplateString)
     }
 
@@ -39,10 +37,6 @@ extension TuistConfig.GenerationOptions {
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
 
-        if container.allKeys.contains(.generateManifest), try container.decodeNil(forKey: .generateManifest) == false {
-            self = .generateManifest
-            return
-        }
         if container.allKeys.contains(.xcodeProjectName), try container.decodeNil(forKey: .xcodeProjectName) == false {
             var associatedValues = try container.nestedUnkeyedContainer(forKey: .xcodeProjectName)
             let templateProjectName = try associatedValues.decode(TemplateString.self)
@@ -56,8 +50,6 @@ extension TuistConfig.GenerationOptions {
         var container = encoder.container(keyedBy: CodingKeys.self)
 
         switch self {
-        case .generateManifest:
-            _ = container.nestedContainer(keyedBy: CodingKeys.self, forKey: .generateManifest)
         case let .xcodeProjectName(templateProjectName):
             var associatedValues = container.nestedUnkeyedContainer(forKey: .xcodeProjectName)
             try associatedValues.encode(templateProjectName)
@@ -72,8 +64,6 @@ public func == (lhs: TuistConfig, rhs: TuistConfig) -> Bool {
 
 public func == (lhs: TuistConfig.GenerationOptions, rhs: TuistConfig.GenerationOptions) -> Bool {
     switch (lhs, rhs) {
-    case (.generateManifest, .generateManifest):
-        return true
     case let (.xcodeProjectName(lhs), .xcodeProjectName(rhs)):
         return lhs.rawString == rhs.rawString
     default: return false