+++ Tests/TuistKitTests/Generator/GeneratorModelLoaderTests.swift
@@ -21,17 +21,14 @@ class GeneratorModelLoaderTest: TuistUnitTestCase {
     typealias RunActionManifest = ProjectDescription.RunAction
     typealias ArgumentsManifest = ProjectDescription.Arguments
 
-    private var manifestTargetGenerator: MockManifestTargetGenerator!
     private var manifestLinter: MockManifestLinter!
 
     override func setUp() {
         super.setUp()
-        manifestTargetGenerator = MockManifestTargetGenerator()
         manifestLinter = MockManifestLinter()
     }
 
     override func tearDown() {
-        manifestTargetGenerator = nil
         manifestLinter = nil
         super.tearDown()
     }
@@ -51,7 +48,6 @@ class GeneratorModelLoaderTest: TuistUnitTestCase {
 
         // Then
         XCTAssertEqual(model.name, "SomeProject")
-        XCTAssertEqual(model.targets.map { $0.name }, ["SomeProject-Manifest"])
     }
 
     func test_loadProject_withTargets() throws {
@@ -75,10 +71,9 @@ class GeneratorModelLoaderTest: TuistUnitTestCase {
         let model = try subject.loadProject(at: temporaryPath)
 
         // Then
-        XCTAssertEqual(model.targets.count, 3)
+        XCTAssertEqual(model.targets.count, 2)
         assert(target: model.targets[0], matches: targetA, at: temporaryPath, generatorPaths: generatorPaths)
         assert(target: model.targets[1], matches: targetB, at: temporaryPath, generatorPaths: generatorPaths)
-        XCTAssertEqual(model.targets[2].name, "Project-Manifest")
     }
 
     func test_loadProject_withManifestTargetOptionDisabled() throws {
@@ -179,8 +174,7 @@ class GeneratorModelLoaderTest: TuistUnitTestCase {
         ]
         let manifestLoader = createManifestLoader(with: manifests, configs: configs)
         let subject = GeneratorModelLoader(manifestLoader: manifestLoader,
-                                           manifestLinter: manifestLinter,
-                                           manifestTargetGenerator: manifestTargetGenerator)
+                                           manifestLinter: manifestLinter)
 
         // When
         let model = try subject.loadProject(at: temporaryPath)
@@ -208,8 +202,7 @@ class GeneratorModelLoaderTest: TuistUnitTestCase {
         ]
         let manifestLoader = createManifestLoader(with: manifests, configs: configs)
         let subject = GeneratorModelLoader(manifestLoader: manifestLoader,
-                                           manifestLinter: manifestLinter,
-                                           manifestTargetGenerator: manifestTargetGenerator)
+                                           manifestLinter: manifestLinter)
 
         // When
         let model = try subject.loadProject(at: temporaryPath)
@@ -680,8 +673,7 @@ class GeneratorModelLoaderTest: TuistUnitTestCase {
 
     func createGeneratorModelLoader(with manifestLoader: GraphManifestLoading) -> GeneratorModelLoader {
         return GeneratorModelLoader(manifestLoader: manifestLoader,
-                                    manifestLinter: manifestLinter,
-                                    manifestTargetGenerator: manifestTargetGenerator)
+                                    manifestLinter: manifestLinter)
     }
 
     func createManifestLoader(with projects: [AbsolutePath: ProjectDescription.Project],